+++ tour-service/src/main/java/com/example/Application.java
@@ -1,13 +1,13 @@
 package com.example;
+import com.example.tourservice.TravelAgentService;
 import com.example.tourservice.utilities.TourInitializer;
 import com.example.tourservice.utilities.TourRepository;
-import com.example.tourservice.utilities.TravelAgentService;
 
 public class Application {
     public static void main(String[] args) {
         TourRepository tr = new TourRepository();
         TourInitializer as = new TourInitializer(tr);
         TravelAgentService agentService = new TravelAgentService(tr);
-        agentService.display();
+        agentService.displayTours();
     }
 }
